fix routes

+++ routes/api.php
@@ -7,9 +7,9 @@ $jobController = new JobController();
 
 $requestUri = explode("?", $_SERVER['REQUEST_URI'])[0];
 
-if (preg_match('/^\/api\/job\/([a-f0-9]{24})$/', $requestUri, $matches) && $_SERVER['REQUEST_METHOD'] === 'DELETE') {
+if (preg_match('/^\/api\/jobs\/([a-f0-9]{24})$/', $requestUri, $matches) && $_SERVER['REQUEST_METHOD'] === 'DELETE') {
     $jobController->deleteJob($matches[1]); // Pasar ID a deleteJob()
-} elseif (preg_match('/^\/api\/job\/([a-f0-9]{24})$/', $requestUri, $matches) && $_SERVER['REQUEST_METHOD'] === 'PUT') {
+} elseif (preg_match('/^\/api\/jobs\/([a-f0-9]{24})$/', $requestUri, $matches) && $_SERVER['REQUEST_METHOD'] === 'PUT') {
     $jobController->updateJob($matches[1]); // Pasar ID a updateJob()
 } else {
     switch ($requestUri) {
@@ -26,13 +26,22 @@ if (preg_match('/^\/api\/job\/([a-f0-9]{24})$/', $requestUri, $matches) && $_SER
             $authController->getUserData();
             break;
         case "/api/jobs":
-            $jobController->getJobsByUser();
+            if ($_SERVER['REQUEST_METHOD'] === 'GET') {
+                $jobController->getJobsByUser();
+            } elseif ($_SERVER['REQUEST_METHOD'] === 'POST') {
+                $jobController->addJob();
+            } else {
+                http_response_code(405);
+                echo json_encode(["error" => "Método no permitido"]);
+            }
             break;
-        case "/api/job/add":
-            $jobController->addJob();
-            break;
-        case "/api/job/:id":
-            $jobController->deleteJob();
+            case (preg_match("#^/api/jobs/\d+$#", $requestUri) ? true : false):
+                if ($_SERVER['REQUEST_METHOD'] === 'DELETE') {
+                    $jobController->deleteJob();
+                } else {
+                    http_response_code(405);
+                    echo json_encode(["error" => "Método no permitido"]);
+                }
             break;
         default:
             http_response_code(404);
